2020年3月7日09:50:56

+++ libr/bin/pdb/pdb_downloader.c
@@ -6,7 +6,7 @@
 #include "pdb_downloader.h"
 
 static bool checkExtract() {
-#if __WINDOWS__ && !__CYGWIN__
+#if __WINDOWS__
 	if (r_sys_cmd ("expand -? >nul") != 0) {
 		return false;
 	}
@@ -31,6 +31,7 @@ static int download(struct SPDBDownloader *pd) {
 	char *curl_cmd = NULL;
 	char *extractor_cmd = NULL;
 	char *abspath_to_archive = NULL;
+	char *abspath_to_file = NULL;
 	char *archive_name = NULL;
 	size_t archive_name_len = 0;
 	char *symbol_store_path = NULL;
@@ -64,13 +65,30 @@ static int download(struct SPDBDownloader *pd) {
 	user_agent = r_str_escape (opt->user_agent);
 	symbol_server = r_str_escape (opt->symbol_server);
 
+	abspath_to_archive = r_str_newf ("%s%s%s%s%s%s%s",
+			    symbol_store_path, R_SYS_DIR,
+			    dbg_file, R_SYS_DIR,
+			    guid, R_SYS_DIR,
+			    archive_name_escaped);
+
+	abspath_to_file = strdup (abspath_to_archive);
+	abspath_to_file[strlen (abspath_to_file) - 1] = 'b';
+	if (r_file_exists (abspath_to_file)) {
+		eprintf ("File already downloaded.\n");
+		R_FREE (user_agent);
+		R_FREE (abspath_to_archive);
+		R_FREE (archive_name_escaped);
+		R_FREE (symbol_store_path);
+		R_FREE (dbg_file);
+		R_FREE (guid);
+		R_FREE (archive_name);
+		R_FREE (abspath_to_file);
+		R_FREE (symbol_server);
+		return 1;
+	}
+
 	if (checkExtract () || opt->extract == 0) {
 		res = 1;
-		abspath_to_archive = r_str_newf ("%s%s%s%s%s%s%s",
-						    symbol_store_path, R_SYS_DIR,
-						    dbg_file, R_SYS_DIR,
-						    guid, R_SYS_DIR,
-						    archive_name_escaped);
 
 		curl_cmd = r_str_newf ("curl -sfLA \"%s\" \"%s/%s/%s/%s\" --create-dirs -o \"%s\"",
 		                       user_agent,
@@ -79,11 +97,9 @@ static int download(struct SPDBDownloader *pd) {
 							   guid,
 		                       archive_name_escaped,
 		                       abspath_to_archive);
-#if __WINDOWS__ && !__CYGWIN__
+#if __WINDOWS__
 		const char *cabextractor = "expand";
 		const char *format = "%s %s %s";
-		char *abspath_to_file = strdup (abspath_to_archive);
-		abspath_to_file[strlen (abspath_to_file) - 1] = 'b';
 
 		// extractor_cmd -> %1 %2 %3
 		// %1 - 'expand'
@@ -101,7 +117,7 @@ static int download(struct SPDBDownloader *pd) {
 		extractor_cmd = r_str_newf (format, cabextractor, abspath_to_dir, abspath_to_archive);
 		R_FREE (abspath_to_dir);
 #endif
-		eprintf ("Attempting to download compressed pdb in %s\n",abspath_to_archive);
+		eprintf ("Attempting to download compressed pdb in %s\n", abspath_to_archive);
 		if ((cmd_ret = r_sys_cmd (curl_cmd) != 0)) {
 			eprintf("curl exited with error %d\n", cmd_ret);
 			res = 0;
@@ -115,19 +131,12 @@ static int download(struct SPDBDownloader *pd) {
 			r_file_rm (abspath_to_archive);
 		}
 		R_FREE (curl_cmd);
-		R_FREE (abspath_to_archive);
 	}
 	if (res == 0) {
 		eprintf ("Falling back to uncompressed pdb\n");
 		res = 1;
-		archive_name[archive_name_len - 1] = 'b';
-		R_FREE (archive_name_escaped);
-		archive_name_escaped = r_str_escape (archive_name);
-		abspath_to_archive = r_str_newf("%s%s%s%s%s%s%s",
-		    symbol_store_path, R_SYS_DIR,
-		    dbg_file, R_SYS_DIR,
-		    guid, R_SYS_DIR,
-		    archive_name_escaped);
+
+		archive_name_escaped[strlen (archive_name_escaped) - 1] = 'b';
 
 		curl_cmd = r_str_newf ("curl -sfLA \"%s\" \"%s/%s/%s/%s\" --create-dirs -o \"%s\"",
 		                       opt->user_agent,
@@ -135,15 +144,16 @@ static int download(struct SPDBDownloader *pd) {
 		                       opt->dbg_file,
 		                       opt->guid,
 		                       archive_name_escaped,
-		                       abspath_to_archive);
-		eprintf ("Attempting to download uncompressed pdb in %s\n",abspath_to_archive);
+		                       abspath_to_file);
+		eprintf ("Attempting to download uncompressed pdb in %s\n", abspath_to_file);
 		if ((cmd_ret = r_sys_cmd (curl_cmd) != 0)) {
 			eprintf("curl exited with error %d\n", cmd_ret);
 			res = 0;
 		}
 		R_FREE (curl_cmd);
-		R_FREE (abspath_to_archive);
 	}
+	R_FREE (abspath_to_archive);
+	R_FREE (abspath_to_file);
 	R_FREE (archive_name);
 	R_FREE (extractor_cmd);
 	R_FREE (symbol_store_path);
@@ -158,6 +168,8 @@ static int download(struct SPDBDownloader *pd) {
 void init_pdb_downloader(SPDBDownloaderOpt *opt, SPDBDownloader *pd) {
 	pd->opt = R_NEW0 (SPDBDownloaderOpt);
 	if (!pd->opt) {
+		pd->download = 0;
+		eprintf ("Cannot allocate memory for SPDBDownloaderOpt.\n");
 		return;
 	}
 	pd->opt->dbg_file = strdup (opt->dbg_file);
@@ -203,7 +215,7 @@ int r_bin_pdb_download(RCore *core, int isradjson, int *actions_done, SPDBOption
 	opt.extract = options->extract;
 
 	init_pdb_downloader (&opt, &pdb_downloader);
-	ret = pdb_downloader.download (&pdb_downloader);
+	ret = pdb_downloader.download ? pdb_downloader.download (&pdb_downloader) : 0;
 	if (isradjson && actions_done) {
 		printf ("%s\"pdb\":{\"file\":\"%s\",\"download\":%s}",
 		        *actions_done ? "," : "", opt.dbg_file, ret ? "true" : "false");
